requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
filelock==3.0.12
Flask==2.0.1
gensim==4.0.1
huggingface-hub==0.0.12
idna==2.10
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.3.4
networkx==2.5.1
nltk==3.6.2
numba==0.53.1
numpy==1.19.5
packaging==20.9
pandas==1.1.5
Pillow==8.2.0
pynndescent==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.5.4
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
sklearn-hierarchical-classification==1.3.2
smart-open==5.1.0
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
transformers==4.8.2
typing-extensions==3.10.0.0
umap-learn==0.5.1
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.4.1
